requirements.txt
rasa-pro
fastapi
uvicorn
PyPDF2
nltk
python-dotenv
openai
pinecone
py-trello

